feat(polyfills): automatically detect and load polyfills as needed
The bootstrapper now automatically loads the es6-shim. It then does
feature detection for html imports and html template elements and loads
them as needed. This fix also hooks up the map config for the framework
and the router so it does not need to be injected when a developer
installs the bootstrapped.