fix(bootstrapper): change how default loader is discovered
This is a BREAKING CHANGE. The bootstrapper now looks for
window.AureliaLoader and uses that to instantiate the loader. If one is
not found, the the presence of the System loader is assumed and the
loader-default is dynamically downloaded and used. By searching for the
global variable, it is now possible to re-use the bootstrapper with
different loaders more easily, provided that they surface the global
AureliaLoader variable and properly shim System.