feat(bootstrapper): new startup logic
This is a BREAKING CHANGE. aurelia-main is no longer supported. Instead
we use aurelia-app for both cases. If not value is provided for the
attr, then we use the default configuration and load app.js as the root
view model. If a value is provided we load that module which serves as
the configuration module. We’ve also added some new apis to aurelia.use
in order to streamline manual configuration.