fix(index): remove html import polyfill load
As a result of refactoring the loader to support different template
loader implementations, this polyfill is no longer loaded by the
bootstrapper. Instead, if you use the html import template loader, it
will load the polyfill, if necessary, when the first template is
requested. In the future that will be moved to a plugin which will load
the polyfill at startup.